+++ CHANGELOG.md
@@ -10,10 +10,10 @@
      - The third parameter is returned if p0 == 0
      - The fourth parameter is returned if p0 > 0
  - Modified effect equations can now use a second parameter, "i", representing total # of iterations
+ - Add x,y,z location equations to Modified effect, for moving the entire effect around.
  - Add Plot effect, mainly intended to be used for testing equations, not really for final effects
  - Fix particle offset x/y/z being an integer
  - ShieldEffect radius changed to floating point
- - Add support for zEquation to Plot effect
 
 # 5.7